# openapi/smart-parking.yaml
openapi: 3.0.3
info:
  title: Smart Parking API
  version: 0.3.0
  description: Step40 integrated contract on top of the Step36-Step37 baseline.
servers:
  - url: http://localhost:8080
    description: gateway
paths:
...
  /api/v1/owner/dashboard:
    get:
      summary: Get owner dashboard view
      parameters:
        - $ref: '#/components/parameters/XTraceId'
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: preferred_window
          in: query
          required: false
          schema:
            type: string
        - name: user_id
          in: query
          required: false
          schema:
            type: string
        - name: order_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Owner dashboard view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnerDashboardResponse'
...
  /api/v1/admin/dashboard:
    get:
      summary: Get admin dashboard view
      parameters:
        - $ref: '#/components/parameters/XTraceId'
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: region_id
          in: query
          required: false
          schema:
            type: string
        - name: trend_days
          in: query
          required: false
          schema:
            type: integer
        - name: trend_limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Admin dashboard view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardResponse'
...
components:
  parameters:
    XTraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
      description: Cross-service trace id.
...
  schemas:
...
    GeoPoint:
      type: object
      required: [lat, lng, display_name]
      properties:
        lat:
          type: number
        lng:
          type: number
        display_name:
          type: string
...
    BillingRule:
      type: object
      required: [timezone, unit_minutes, rounding_mode, effective_window]
      properties:
        timezone:
          type: string
        unit_minutes:
          type: integer
        rounding_mode:
          type: string
        effective_window:
          type: string
    BillingRecord:
      type: object
      required: [order_id, reservation_id, slot_id, region_id, billable_minutes, estimated_amount, final_amount, billing_status]
      properties:
        order_id:
          type: string
        reservation_id:
          type: string
        user_id:
          type: string
        slot_id:
          type: string
        region_id:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        billable_minutes:
          type: integer
        estimated_amount:
          type: number
        final_amount:
          type: number
        billing_status:
          type: string
        recognized_on:
          type: string
    RegionRevenueSummary:
      type: object
      required: [date, region_id, revenue_amount, order_count, utilization_rate]
      properties:
        date:
          type: string
        region_id:
          type: string
        revenue_amount:
          type: number
        order_count:
          type: integer
        utilization_rate:
          type: number
    RecommendationItem:
      type: object
      required: [slot_id, estimated_amount, eta_minutes, destination, map_url]
      properties:
        slot_id:
          type: string
        estimated_amount:
          type: number
        eta_minutes:
          type: integer
        destination:
          $ref: '#/components/schemas/GeoPoint'
        map_url:
          type: string
        slot_display_name:
          type: string
        region_label:
          type: string
...
    OwnerOrderDetail:
      allOf:
        - $ref: '#/components/schemas/BillingRecord'
        - type: object
          properties:
            billing_rule:
              $ref: '#/components/schemas/BillingRule'
...
    OwnerJourneyState:
      type: object
      required: [state, message]
      properties:
        state:
          type: string
        message:
          type: string
    OwnerDashboardSummary:
      type: object
      required: [region_id, region_label, recommendation_count, latest_order_id, latest_billing_status, latest_amount]
      properties:
        region_id:
          type: string
        region_label:
          type: string
        recommendation_count:
          type: integer
        latest_order_id:
          type: string
        latest_billing_status:
          type: string
        latest_amount:
          type: number
    OwnerDashboardResponse:
      type: object
      required: [location, preferred_window, user_id, summary, journey, billing_rule, recommendations, trace_id, service]
      properties:
        location:
          type: string
        preferred_window:
          type: string
        user_id:
          type: string
        summary:
          $ref: '#/components/schemas/OwnerDashboardSummary'
        journey:
          $ref: '#/components/schemas/OwnerJourneyState'
        billing_rule:
          $ref: '#/components/schemas/BillingRule'
        latest_order:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/OwnerOrderDetail'
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/RecommendationItem'
        trace_id:
          type: string
        service:
          type: string
    AdminDashboardSummary:
      type: object
      required: [date, region_id, active_reservations, occupancy_rate, revenue_total, dispatch_strategy, business_view]
      properties:
        date:
          type: string
        region_id:
          type: string
        active_reservations:
          type: integer
        occupancy_rate:
          type: number
        revenue_total:
          type: number
        dispatch_strategy:
          type: string
        business_view:
          type: boolean
    AdminDashboardHighlights:
      type: object
      required: [region_count, revenue_points, forecast_points, peak_occupancy]
      properties:
        region_count:
          type: integer
        revenue_points:
          type: integer
        forecast_points:
          type: integer
        peak_occupancy:
          type: number
    AdminDashboardSections:
      type: object
      required: [revenue_summary, revenue_trend, occupancy_trend, forecast_compare]
      properties:
        revenue_summary:
          type: array
          items:
            $ref: '#/components/schemas/RegionRevenueSummary'
        revenue_trend:
          type: array
          items:
            $ref: '#/components/schemas/RevenueTrendPoint'
        occupancy_trend:
          type: array
          items:
            $ref: '#/components/schemas/OccupancyTrendPoint'
        forecast_compare:
          type: array
          items:
            $ref: '#/components/schemas/ForecastComparePoint'
    DegradedMetadata:
      type: object
      required: [realtime_transport, chart_mode, data_sources]
      properties:
        realtime_transport:
          type: string
        chart_mode:
          type: string
        data_sources:
          type: array
          items:
            type: string
    AdminDashboardResponse:
      type: object
      required: [summary, highlights, sections, diagnostic_links, degraded_metadata, trace_id, service]
      properties:
        summary:
          $ref: '#/components/schemas/AdminDashboardSummary'
        highlights:
          $ref: '#/components/schemas/AdminDashboardHighlights'
        sections:
          $ref: '#/components/schemas/AdminDashboardSections'
        diagnostic_links:
          type: object
          additionalProperties:
            type: string
        degraded_metadata:
          $ref: '#/components/schemas/DegradedMetadata'
        trace_id:
          type: string
        service:
          type: string
...
    RevenueTrendPoint:
      type: object
      required: [date, revenue_amount, order_count]
      properties:
        date:
          type: string
        revenue_amount:
          type: number
        order_count:
          type: integer
...
    OccupancyTrendPoint:
      type: object
      required: [ts, region_id, occupancy_rate, active_supply]
      properties:
        ts:
          type: string
        region_id:
          type: string
        occupancy_rate:
          type: number
        active_supply:
          type: number
...
    ForecastComparePoint:
      type: object
      required: [ts, region_id, predicted_gap, actual_gap]
      properties:
        ts:
          type: string
        region_id:
          type: string
        predicted_gap:
          type: number
        actual_gap:
          type: number
...